pyproject.toml
[tool.poetry]
name = "what-i-should-learn"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "what_i_should_learn"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
scrapy = "^2.10.1"
cloudscraper = "^1.2.71"
scrapeops-scrapy-proxy-sdk = "^1.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
